Move dos location tests to new file and pick up postaltown from table
